requirements-scraper.txt
beautifulsoup4
habanero
requests
ruamel.yaml
